Refactoring of hook behaviour of MapPage
