+++ src/frontend/config.jsx
@@ -3,3 +3,6 @@ export const MAX_COVERED_MONTHS = 6
 // Labels for visualization
 export const MONTH_LABELS = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec']
 export const DAY_ORDER = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday']
+// Limit the number of stations to fetch for performance reasons
+export const LIMIT_STATIONS = 3000
+export const LIMIT_TRIPS = 1000
