+++ src/frontend/map/hooks/useTripFlowLayer.jsx
@@ -1,0 +1,30 @@
+import { useMemo } from 'react'
+import { selectTrips, selectMaxFlow } from '../selectors/tripFlow.js'
+import useTripCounts from '../../hooks/useTripCounts.js'
+import { LIMIT_TRIPS } from '../../config.jsx'
+
+/**
+ * Custom hook to fetch and process trip flow data for the trip flow layer.
+ * It retrieves trip counts, filters them based on the date range, and calculates the maximum flow for scaling purposes.
+ * @param {Object} dateRange - Optional date range for filtering trip flow data.
+ * @returns {Object} An object containing the filtered trip data, maximum flow value, loading state, and error state.
+ */
+export function useTripFlowLayer({ dateRange }) {
+    // Build filters for trip count data
+    const tripCountFilters = {
+        limit: LIMIT_TRIPS,
+        ...(dateRange ?? {})
+    }
+
+    // Fetch trip count data for the current date range.
+    const { tripCount,
+        loading: loading,
+        error: error
+    } = useTripCounts(tripCountFilters)
+
+    // Process trip data to get the trips and calculate the maximum flow for scaling
+    const trips = useMemo(() => selectTrips(tripCount), [tripCount])
+    const maxTripFlow = useMemo(() => selectMaxFlow(tripCount), [tripCount])
+
+    return { trips, maxTripFlow, loading, error}
+}
