+++ src/frontend/map/layers/tripFlowLayer.jsx
@@ -1,5 +1,6 @@
 import { ArcLayer } from '@deck.gl/layers'
 
+// Constants for arc styling based on trip usage
 const BASE_ALPHA = 80
 const MAX_ALPHA_RANGE = 175
 const BASE_WIDTH = 1
@@ -39,8 +40,8 @@ function getArcColor(baseColor, normalizedUsage) {
 
 /**
  * Creates a layer for displaying frequent trips based on their usage
- * @param {Array} params.trips - Array of trip objects with sourcePosition, targetPosition, and dailyFlow
- * @param {number} params.maxTripCount - Maximum trip count for scaling widths and colors
+ * @param {Array} trips - Array of trip objects with sourcePosition, targetPosition, and dailyFlow
+ * @param {number} maxTripCount - Maximum trip count for scaling widths and colors
  * @returns {ArcLayer}
  */
 export function createTripFlowLayer({ trips, maxTripCount }) {
@@ -71,14 +72,26 @@ export function createTripFlowLayer({ trips, maxTripCount }) {
 
 
 /**
- * Generates a tooltip for trip flow data, showing the number of rides between two stations.
- * @param {Object} object - The trip flow data object.
- * @returns {string} The tooltip content.
- */
+* Generates a tooltip for trip flow data, showing the number of rides between two stations.
+* @param {Object} object - The trip flow data object.
+* @returns {string} The tooltip content.
+*/
 export function tripFlowTooltip(object) {
     console.log('Generating tooltip for trip flow:', object)
     const rides = Math.round(Number(object.total_daily_flow) || 0)
-    const from = object.start_station_name 
-    const to = object.end_station_name 
+    const from = object.start_station_name
+    const to = object.end_station_name
     return `Trip: ${from} → ${to}\nRides: ${rides}`
+}
+
+/**
+* Creates a legend for the trip flow layer, indicating that the arcs represent frequent trips.
+* @returns {JSX.Element} The legend component for the trip flow layer.
+*/
+export function tripFlowLegend() {
+    return (
+        <div className="map-legend">
+            <p className="map-legend-text">Frequent trips</p>
+        </div>
+    )
 }
